requirements.txt
flask
openai
python-decouple